Time: 144 ms (65.26%), Space: 71.4 MB (28.67%) - LeetHub

+++ 2439-minimize-maximum-of-array/2439-minimize-maximum-of-array.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int minimizeArrayValue(vector<int>& nums) {
-        long sum = 0, res = 0;
+        long long sum = 0, res = 0;
         for (int i = 0; i < nums.size(); ++i) {
             sum += nums[i];
-            res = max(res, (long)ceil((double)sum/(i+1)));
+            res = max(res, (long long)ceil((double)sum/(i+1)));
         }
         return res;
     }
